chore: remove issue template in prep for open-sourcing
This repo's issue template was carried over from an
internal repo. It contains placeholder values for GitHub
handles that are intended to be replaced by real ones,
but those handles already are real users here on GitHub.com,
so we'd be spamming them with each new issues if we were
to keep this in.

In addition, this template isn't used in any of our other soon-
to-be open-sourced repos, so it feels like an oddball. This PR
removes it altogether for now. We may re-add a template in the
future, but it will be consistent with other repos.