Updated to use PostgresDSN.build. This required the addtion of INST_DB_SCHEME, which I've updated in the .env files, but will write stories to update the AWS secrets.
Removed the INST_DB_CONN from the env. If we want to keep this, I can add an if in the
field_validator to check for existence, but that makes encoding the pwd tricky again.  Unless
we assume if the INST_DB_CONN exists, that it's local dev and ignore pwd encoding.